RdmaChannel: Added a "mustRetry" flag to getRdmaChannel()
Some getRdmaChannel() calls are considered "best effort": in the case of sending a helloMsg to
the driver or the driver sending the announceMsg to executors. In this case, try to create
the connection once, and fail silently if an error occurs.

If this connection is required later, getRdmaChannel() will be called again and errors will
be handled appropriately.

Change-Id: I3350b90f4942e3bf648fd26a8e2fca3b801ed037
Signed-off-by: Vu Pham <vu@mellanox.com>